chore: move error hook beside form hook
